--- docs/reference/API/trestle/transforms/results.md
@@ -1,7 +1,0 @@
----
-title: trestle.transforms.results
-description: Documentation for trestle.transforms.results module
----
-
-::: trestle.transforms.results
-handler: python
