--- docs/reference/API/trestle/transforms/transformer_factory.md
@@ -1,7 +1,0 @@
----
-title: trestle.transforms.transformer_factory
-description: Documentation for trestle.transforms.transformer_factory module
----
-
-::: trestle.transforms.transformer_factory
-handler: python
