--- docs/reference/API/trestle/transforms/transformer_helper.md
@@ -1,7 +1,0 @@
----
-title: trestle.transforms.transformer_helper
-description: Documentation for trestle.transforms.transformer_helper module
----
-
-::: trestle.transforms.transformer_helper
-handler: python
